html/interaction/focus/sequential-focus-navigation-and-the-tabindex-attribute/focus-tabindex-order-shadow-zero-slot-one.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: focus - the sequential focus navigation order with shadow dom and all tabindex=0, except slot (tabindex=1)</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/interaction.html#sequential-focus-navigation">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<body>
<div id="aboveHost">aboveHost</div>
<div id="host">
  <div id="slotted">slotted</div>
</div>
<div id="belowHost">belowHost</div>
</body>

<script>
const shadowRoot = host.attachShadow({ mode: "open" });
const aboveSlot = document.createElement("div");
aboveSlot.innerText = "aboveSlot";
const slot = document.createElement("slot");
const belowSlot = document.createElement("div");
belowSlot.innerText = "belowSlot";
shadowRoot.appendChild(aboveSlot);
shadowRoot.appendChild(slot);
shadowRoot.appendChild(belowSlot);

// Structure:
// <div #aboveHost tabindex=0>
// <div #host tabindex=0>
//    #shadowRoot
//      <div #aboveSlot tabindex=0>
//      <slot #slot tabindex=1>
//        (slotted) <div #slotted tabindex=0>
//      <div #belowSlot tabindex=0>
// <div #belowHost tabindex=0>

const elementsInFlatTreeOrder = [aboveHost, host, aboveSlot, slot, slotted, belowSlot, belowHost];

function setAllTabIndex(value) {
  for (let el of elementsInFlatTreeOrder) {
    if (el == slot)
      el.tabIndex = 1;
    else
      el.tabIndex = 0;
  }
}

function navigateFocusForward() {
  return new Promise((resolve, reject) => {
    // TAB = '\ue004'
    test_driver.send_keys(document.body, "\ue004").then(() => {
      if (shadowRoot.activeElement)
        resolve(shadowRoot.activeElement);
      resolve(document.activeElement);
    });
  });
}

async_test(async (t) => {
  setAllTabIndex();
  document.body.focus();
  // Focus should move first according to flat tree order to #aboveHost and #host, then into #host's focus scope.
  // It will then move to #slotted because #slot has tabindex=1 (though we actually won't focus on slot),
  // and then back to #host's focus scope again, finally getting out to the document focus scope.
  const expectedOrder = [aboveHost, host, slotted, aboveSlot, belowSlot, belowHost];
  for (let el of expectedOrder) {
    let focused = await navigateFocusForward();
    t.step(() => { assert_equals(focused, el) });
  }
  t.done();
}, "Order when all tabindex=0 except for slot, which has tabindex=1");
</script>
